split main in smaller packages
